enemy_estimatorをestimatorに統合
