+++ consai2_world_observer/include/world_observer/estimator.hpp
@@ -51,5 +51,42 @@ class Estimator
 };
 
 
+class EnemyEstimator : public Estimator
+{
+    public:
+        // constructor
+        EnemyEstimator();
+
+        void Init(double loop_time);
+
+        ~EnemyEstimator();
+
+    protected:
+        double  dt;
+
+        void initSystemModel();
+        void initMeasurementModel();
+
+        void  predict(ColumnVector input);
+        void  update(ColumnVector measurement);
+
+        geometry2d::Odometry convetEstimationToOdometry();
+
+        bool isOutlier(ColumnVector measurement);
+        double mahalanobisDistance(ColumnVector measurement);
+
+        Estimation getResult();
+
+        void collectAngleOverflow(ColumnVector& state, SymmetricMatrix& cov);
+};
+
+
+
+class EulerAngle
+{
+    public:
+        static double  normalize(double angle);
+        static double  normalize(double angle, double center);
+};
 
 #endif
