+++ consai2_world_observer/src/main.cpp
@@ -4,7 +4,6 @@
 
 #include <world_observer/world_observer_ros.hpp>
 #include <world_observer/estimator.hpp>
-#include <world_observer/enemy_estimator.hpp>
 
 // AppearanceMonitor　クラス
 // ロボットまたはボールの出現、および消失をモニタリングするクラス
